Support fractional lead automation hours

+++ Models/LeadAutomationSettings.cs
@@ -3,7 +3,7 @@ namespace backend.Models;
 public class LeadAutomationSettings
 {
     public int Id { get; set; } = 1;
-    public int UnassignAfterHours { get; set; } = 24;
-    public int ReminderIntervalHours { get; set; } = 1;
+    public decimal UnassignAfterHours { get; set; } = 24;
+    public decimal ReminderIntervalHours { get; set; } = 1;
     public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
 }
